requirements.txt
torch==1.13.1
torchvision==0.14.1
numpy==1.23.5
Pillow==9.4.0
tensorboard==2.12.0
six==1.16.0
tqdm==4.66.4
urllib3~=2.5.0